Add gameBoard module
The gameBoard object stores the tic tac toe board.
It has the reset method for clearing the board,
the move method for making a game move, The update
method for updating the DOM with the current game
state, checkWin for checking if a given player has
won, checkFull for checking if the board is full.